requirements.txt
torch>=2.0
torchvision>=0.15
python-chess>=1.7
PyYAML>=6.0
tensorboard>=2.12
wandb>=0.15
numpy>=1.24
# optional: torch-compile

